Allow ci-chat-bot to teardown launched clusters
The difference between edit and admin permissions is the ability to delete
the project and the ability to add other users. The former is required to
allow us to let clusters last longer (we have to allow releasing the cluster
when it's no longer needed), and the later may be useful in the future.

There should be no risk to this change from a security perspective.