templates/openshift: add openshift-install invoker
OpenShift is going to soon begin reporting information about the
installation process via telemetry. This will include the
openshift-install version and the invoker (usually the user that invoked
the program). In order to help distinguish between our customers'
clusters and the ones created by CI, the invoker is overridden with a
fairly unique and recognizable name (internal-openshift-ci/${JOB_NAME}).