ci-operator/jobs/openshift/installer: Report pull-ci-openshift-installer-master-e2e
The current approach is a minimum viable implementation (there's a
more scalable implementation in the works).  And non-AWS tests usually
fail.  But neither of those are reasons to not report this job to
GitHub.  We want to work on getting the other platforms passing, so
it's useful to have easy links to their results for a given PR.  And
installer developers don't really care about the platform fuzzing
implementation; they'll just continue to consume the results as the
implementation improves.