Get real status of e2e-openstack install task
Previously the script checked the status of the last command, which
happened to be the invocation of the installer.  With
https://github.com/openshift/release/commit/5fa33bd7ad2e12090c7b2824e057b2f0030b07cb
we've changed it to make an HTTP request at the end of it. This caused
the script to wrongly return success even if the installer failed.

This also removes the unused baseImage option which has been removed
from the installer.