[DNM] [Test Only] Get real status of e2e-openstack install task
Previously the script checked the status of the last command, which
happened to be the invocation of the installer.  With
https://github.com/openshift/release/commit/ede673b1d5ee989b55b90d57401471d503ca065c
we've changed it to make an HTTP request at the end of it. This caused
the script to wrongly return success even if the installer failed.

This also removes the unused baseImage option which has been removed
from the installer.