openstack: drop bootstrap_ip log code from CI templates
OpenStack CI jobs only target 4.2. As such the preference is to rely on
'gather bootstrap' to obtain the CI bootstrap logs. The code here won't
work because it does not have the correct logic to obtain the bootstrap
node IP anyway.

This patch removes the inline code copied from an earlier version of the
example templates.