CI configuration for openshift/coredns-mdns
Depends on https://github.com/openshift/coredns-mdns/pull/39

This only adds a go fmt test to the repo. The repo is a plugin for
coredns so it doesn't become a standalone image, and the go vet tests
are currently failing on it. If/when we get those working we can add
them as well.