Bump the limit of rehearsed jobs
We are not rehearsing any jobs for repositories with many presubmits,
which is the opposite of what we want to be doing. In general we should
not worry about starting a thundering herd if all jobs belong to one
configuration, as any new pull request to the same repository will
trigger the same set of jobs normally.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>