Remove OpenStack beta RPM repos for 4.2
The side-effect of this is that no 4.2 images will
ever be built with the correct RPMs as the the original
configuration that was supplied in the upstream image
build process pointed to the prior release's software.

Performing as a separate change so the prior configuration
can be recovered and reviewed as opposed to force-push
overwrite.