config-brancher --current-release=4.2 --future-release=4.3
This time, the sync is manually tweaked:
1. https://github.com/openshift/ci-tools/pull/142 version of
config-brancher was used, because `config-brancher` from the current
HEAD was segfaulting
2. I omitted the sync of `openshift/cluster-node-tuning-operator` and
`openshift/openshift-tuned`, because a previous
https://github.com/openshift/release/pull/4976 PR mistakenly only
modified the 4.2 and 4.3 versions of the configs, so the sync would
regress their behavior on those branches. I'll file a followup PR to
include those changes also in the master config.