templates/openshift: Fail when 'destroy cluster' fails
The 'teardown' containers currently do both log-collection and cluster
teardown.  We probably want to split those into two separate
containers, because they are so different.  But currently we don't
'set -e' those containers because many of the log-gathering commands
are oportunistic and we accept their failure.  Similarly in the AWS
UPI template, we call 'aws cloudformation delete-stack' on all
potential stacks, even if the 'setup' container died before getting
through them all, and that means delete-stack will fail when called
with the later, never-created stack names.

But 'destroy cluster' should always succeed, or we want to hear about
it.  So with this commit its errors become fatal for the 'teardown'
container.