core-services/prow/02_config/_boskos: Bump AWS leases to 150
The 120 from 9507031510 (Add quota leasing to all jobs, 2019-08-09, #4699)
was motivated by us running peak loads of around 125 concurrent AWS
errors and seeing errors like:

  level=error msg="Error: timeout while waiting for state to become 'active' (last state: 'provisioning', timeout: 10m0s)"

But that's an AWS performance error (and sometimes a throttling
error), not an actual resource limit [1], and our current leasing
strategy is more geared towards resource capacity (or at least the
slippery assumption that AWS API request throughput is roughly
constant over the life of the test).  So it's hard to put an exact
number on what our limit should be.  We're currently riding near the
120 cap for hours at a time during our peak-use periods, so this
commit bumps that up a bit so we can get a better feel on possible
headroom.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1717604#c20