Fix err msg in validate-owners.sh
ci-operator/populate-owners.sh is removed.

The new tool openshift/ci-tools/autoowners for the moment does not
work directly for one directory and requires the user to have a
token to have access to all repos.

In order for users to pass the OWNERS check, when adding new repos,
we hint the user to copy the OWNERS file over from the upstream.

Any changes on the OWNERS or OWNERS_ALIASES  afterwards will be
synced to release repo automatically by the ci-tool.