openstack: mount the etcd storage in a ramdisk
The OpenStack cloud our CI uses suffers from unpredictable disk
latency spikes (and the baseline performance isn't that great either).
This is causing a lot of spurious test failures because etcd goes down
randomly and brings the OpenShift API with it.

By putting the etcd storage in a ramdisk, the latency is consistently
under 1ms making the tests more reliable.

We do not want to ship this change in the installer/MCO because it
makes the etcd data non-persistent. This is fine for CI where the
cluster is destroyed an hour later anyway, but not for production
environments.