OpenStack: Add note about boskos.ci
The `e2e-openstack` pod uses a custom resolver hosted in the OpenStack
CI infrastructure on MOC, in order to resolve the `api` and `*.apps`
names for the cluster under test. This custom resolver needs to know
the IP address of the `boskos.ci` service running in the CI cluster, so
document it here.