ci-operator/templates/openshift/installer/cluster-launch-installer-e2e: Don't update OPENSHIFT_INSTALL_RELEASE_IMAGE_OVERRIDE for mirrors
This is from 246118f38e (Add support for testing installs from
mirrored repositories, 2019-07-22, #4585), because there was some
confusion about how the bootstrap machine would decide to pull from
the mirror instead of the original.  But Clayton points out that that
the imageContentSources must apply to the bootstrap machine as well,
so we can use the original release pullspec and the bootstrap machine
will still pull it from the mirror.