Allow passing a github token for making jobs
When we need to create a new run of a job that requires us to look up
the current state of a branch or ref, we need to use a GitHub client.
Often the anonymous client used by default will be rate-limited so we
should just allow users to pass in their own.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>