Make Tripleo based zuul jobs non-blocking on libpod PRs
TripleO, an Openstack deployment tool is one of the consumer of
podman to run OpenStack services in container. In order to avoid
breakage due to podman releases, TripleO based Zuul jobs are
getting added.

But libpod PRs are getting blocked due to zuul jobs failure.
In order to fix that, we are setting skip-unknown-contexts
to true makes the tide to ignore the zuul job result.

Signed-off-by: Chandan Kumar (raukadah) <raukadah@gmail.com>