ci-operator/templates: Pull AWS region and zone from the cluster for TEST_PROVIDER
This isn't as nice as having the suite extract them for us, but it's
something we can do in openshift/release vs. openshift/origin ;).  We
could get both the region and zone by inspecting node labels, but the
infrastructure config seems like a more canonical location for the
region information, so I'm pulling it from there.  I don't feel
particularly strongly about it, though, and would be fine pivoting to
pull it from a node like we pull the zone.  We could also check nodes
to determine multizone-ness and multimaster-ness, but I'm leaving
those hard-coded for now.