ci-operator/templates/openshift/installer: Bump AWS UPI to RHCOS 43.81.201911221453.0
Catching up with openshift/installer@a8e9991876 (rhcos: Bump to
43.81.201911221453.0, 2019-11-22, openshift/installer#2714) and its
encryption fixes.

Unfortunately we still don't have a way to expose the installer's
recommended default, and I'm still opposed to reaching into the
installer sources and looking in its internal rhcos.json like
c85599a25b (templates/installer/e2e-metal: use the rhcos.json from
upi-installer image to pick bootimage, 2019-04-26, #3612) is doing for
metal, because that's not what our users will be doing.  Using the
installer's tip AMI defaults is also not what users will be doing, but
I'm not clear on how we can select the right branch in the templates
to pull from the openshift-docs defaults [1,2,3] (which is what users
will be doing).

[1]: https://docs.openshift.com/container-platform/4.1/installing/installing_aws_user_infra/installing-aws-user-infra.html#installation-aws-user-infra-rhcos-ami_installing-aws-user-infra
[2]: https://docs.openshift.com/container-platform/4.2/installing/installing_aws_user_infra/installing-aws-user-infra.html#installation-aws-user-infra-rhcos-ami_installing-aws-user-infra
[3]: https://github.com/openshift/openshift-docs/blob/b24fe4684f8f6ba11cae92c5a779abbe4af57717/modules/installation-aws-user-infra-rhcos-ami.adoc#op-system-amis-for-the-aws-infrastructure