Enforce vendoring mode for OpenShift org
This has always been the case. Modules are on by default with 1.13
and until modules are fixed, they ignore vendor folder,
unless explicitily asked to use vendoring mode. This change ensures
that using vendor folder is default choice. We have always done
it and this codifies that practice. People can opt out by specifying the
flag manually on go invocation but it is discouraged. The whole org
should vendor its dependencies to allow for offline and reproducible
builds.