SDCICD-56. Remove dependency management from openshift/release.
The openshift/release project will no longer perform the glide
dependency installation manually. This will allow the osde2e repo to
control how it handles dependencies. This is preliminary to switching
from glide to godeps, with the eventual hope that we'll use the
Dockerfile in the osde2e repo to run jobs.