ci-operator/templates/openshift/installer/cluster-launch-installer-upi-src: Drop unused template
The template was added in 78d8c17337 (Add a UPI based src template,
2019-10-11, #5388), but had no consumers.  There were potential
consumers in flight with [1,2], but both of those pull requests closed
without landing.  We can always dig this back out of the Git history
if it turns out folks actually do need a non-e2e UPI template, but
until then remove it to save the maintenance cost of keeping unused
code up to date.

[1]: https://github.com/openshift/release/pull/4325/files#diff-264b567e579413709feec76ef0604cf3R336
[2]: https://github.com/openshift/release/pull/5367/files#diff-131c135812b5130dcd1252473291b4a5R188