ci-operator/templates/openshift/installer/cluster-launch-installer-upi*: Switch on BRANCH for RHCOS AMI
As discussed in dc26075230 (ci-operator/templates/openshift/installer:
Bump AWS UPI to RHCOS 43.81.201911221453.0, 2019-12-02, #6192), we
want a way to switch on the z stream when selecting the RHCOS AMI.  We
amight also need this if other parts of the recommended
user-provided-infrastructure flow diverge between z streams, although
at the moment the AMI is the only divergence.

In this commit, I'm sourcing my AMIs with:

  $ git cat-file -p origin/release-4.1:data/data/rhcos.json | jq -r '.amis["us-east-1"].hvm'
  ami-046fe691f52a953f9
  $ git cat-file -p origin/release-4.2:data/data/rhcos.json | jq -r '.amis["us-east-1"].hvm'
  ami-01e7fdcb66157b224
  $ git cat-file -p origin/release-4.3:data/data/rhcos.json | jq -r '.amis["us-east-1"].hvm'
  ami-014ce8846db8b463d
  $ git cat-file -p origin/master:data/data/rhcos.json | jq -r '.amis["us-east-1"].hvm'
  ami-014ce8846db8b463d

as the installer stood at 2019-12-04T23:30Z.  Comparing with the user-facing docs:

  $ git --no-pager grep -hA1 us-east-1 origin/enterprise-4.1:modules/installation-aws-user-infra-rhcos-ami.adoc
  |`us-east-1`
  |`ami-046fe691f52a953f9`
  $ git --no-pager grep -hA1 us-east-1 origin/enterprise-4.2:modules/installation-aws-user-infra-rhcos-ami.adoc
  |`us-east-1`
  |`ami-01e7fdcb66157b224`
  $ git --no-pager grep -hA1 us-east-1 origin/enterprise-4.3:modules/installation-aws-user-infra-rhcos-ami.adoc
  |`us-east-1`
  |`ami-01e7fdcb66157b224`
  $ git --no-pager grep -hA1 us-east-1 origin/enterprise-4.4:modules/installation-aws-user-infra-rhcos-ami.adoc
  fatal: Invalid object name 'origin/enterprise-4.4'.
  $ git --no-pager grep -hA1 us-east-1 origin/master:modules/installation-aws-user-infra-rhcos-ami.adoc
  |`us-east-1`
  |`ami-01e7fdcb66157b224`

So we currently diverge for 4.3 and master, where I have [1,2] open to
close the gaps.

The BRANCH environment variable has been around since 22e30eac2b (The
release promotion periodic can be conditional and no-op if no change,
2019-04-30, #3649).  I added it to all jobs based on the UPI template
where it was missing (it was already on a number of those jobs).

[1]: https://github.com/openshift/openshift-docs/pull/18396
[2]: https://github.com/openshift/openshift-docs/pull/18395