Correct ClusterAutoscaler set up on build01
`ClusterAutoscaler.spec.cores.min/max` expects the total numbers
of cores of ALL workers.

We want 3 - 12 workers to scale up/down and each worker has 16
cores as `m5.4xlarge` instance.

The old config was set in the wrong way that it is about the
number of cores for EACH workers to deploy. We fix it, together
with memory, with this PR.

https://coreos.slack.com/archives/CBZHF4DHC/p1578326297014800