Lease handling for conformance tests suite
To better utilize our resources we determined that
run-minimal test will create a cluster with half
the resources needed for conformance.
This patch introduces the logic for acquiring 2 leases
for conformance and one for minimal suite.
Notice that at the moment since we have separate ovirt clusters
we needed to make sure that the first and second lease will use
the same ovirt cluster.

Signed-off-by: Gal Zaidman <gzaidman@redhat.com>