step-registry: add configure and install IPI steps
Essentially a copy of the `setup` container from the templates broken into a
configuration and an install step.  Communication between the two is done by
writing/reading `install-config.yaml` to the shared secret.

The configuration step checks the cluster type and special files in the shared
secret to customize the installation.