installer templates to limit destroy calls to 30 minutes
Prior to this change, destroy cluster would occasionally get suck in a
loop on the gcp platform. This caused ci to reach max time and never
collect logs from the failed teardown.

This change uses the timeout command to limit these calls to 30m. As a
result, the failure should be captured in the logs and the logs properly
added to the archive.