Add ports for four more clusters to the Z bastion
We want to try running more than one cluster simultaneously, so we need
to forward the traffic for each of the cluster slots. The system
currently chooses a slot based on the job hash (effectively randomly),
so we need enough slots to ensure that the odds are decent that there
won't be a collision. This is a temporary hack and we'll want something
more robust in the future.

The SSH invocation on the other end looks like the following:

-R 16509:127.0.0.1:16509 \
-R 6443:192.168.126.11:6443 -R 8080:192.168.126.51:80 -R 8443:192.168.126.51:443 \
-R 2210:192.168.126.10:22 -R 2211:192.168.126.11:22 -R 2212:192.168.126.12:22 -R 2213:192.168.126.13:22 \
-R 16443:192.168.1.11:6443 -R 10080:192.168.1.51:80 -R 10443:192.168.1.51:443 \
-R 26443:192.168.2.11:6443 -R 20080:192.168.2.51:80 -R 20443:192.168.2.51:443 \
-R 36443:192.168.3.11:6443 -R 30080:192.168.3.51:80 -R 30443:192.168.3.51:443 \
-R 46443:192.168.4.11:6443 -R 40080:192.168.4.51:80 -R 40443:192.168.4.51:443 \