VSphere UPI: Add cluster to install-config.
As a requirement for VSphere IPI the cluster field will be required in the install-config. Adding the field here would prevent breaking the UPI test because the install config would be missing the field. This should have no effect on the UPI test. Note that the cluster is already being provided in the terraform variables.