OpenStack: reduce to number of simultaneous jobs
The cloud on which OpenStack CI runs is having a long standing capacity
issue impacting our jobs. Many VMs (bootstrap, masters or workers) end
up in ERROR state and can be difficult to diagnose if you're not aware
of the problem [1].

Let's reduce the number of simultaneous possible jobs from 10 to 3 to
see if this has a meaningful impact on our pass rate.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1788316