ci-operator/config/openshift/cincinnati/openshift-cincinnati-master: Check Cargo.lock, etc.
Require Cargo.lock to be fresh when we build.  Not sure if there's a
way to run this to only bump when required, but we generally want to
trust our dependencies to correctly SemVer and track their upstream
bugfixes, so I'm not too concerned about this failing after a
dependency releases a new patch version or whatever.

We don't need &&-chaining or anything, because this eventually gets
called with errexit.  For example, [1]:

  --> RUN ["/bin/bash","-c","set -o errexit; umask 0002; echo \"Checking Cargo.lock\" \u0026\u0026\ncargo generate-lockfile \u0026\u0026\ngit diff --exit-code \u0026\u0026\necho \"Building release binaries\" \u0026\u0026\ncargo build --locked --release\n"]

Also spit out the Rust and Cargo versions to make it easier to
understand why they recommend lockfile changes when they do recommend
changes.

The codegen-protoc business follows [2] and, with the 'git diff',
ensures the automatically-generated
cincinnati/src/plugins/interface.rs is up to date too.  The cd
business avoids [3]:

  error: --features is not allowed in the root of a virtual workspace

from:

  cargo build --features=codegen-protoc --release

and [4]:

  error: the `-Z` flag is only accepted on the nightly channel of Cargo, but this is the `stable` channel

from:

  cargo build -Z package-features --features=codegen-protoc --release -p cincinnati

because they were building in a root outside of the
cincinnati/Cargo.toml package that actually needs the feature flag
[5].

[1]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_release/6935/rehearse-6935-pull-ci-openshift-cincinnati-master-images/4/build-log.txt
[2]: https://github.com/openshift/cincinnati/tree/a03ccf39ce8633f8585272053bc595ac28fe9ed0#updating-the-plugin-interface-scheme
[3]: https://prow.svc.ci.openshift.org/view/gcs/origin-ci-test/pr-logs/pull/openshift_release/6935/rehearse-6935-pull-ci-openshift-cincinnati-master-images/8#1:build-log.txt%3A30
[4]: https://prow.svc.ci.openshift.org/view/gcs/origin-ci-test/pr-logs/pull/openshift_release/6935/rehearse-6935-pull-ci-openshift-cincinnati-master-images/10#1:build-log.txt%3A30
[5]: https://github.com/rust-lang/cargo/pull/5353