update periodic job configuration to allow reherse
The periodic job did not successfully bring up a cluster and failed
on 7 consecutive occasions for the following error. The problem is
that the job configuration did not use `JOB_NAME_SAFE` for cluster
profile name which is required. The Make jobs from openshift/release
did not point out this problem and consistently deletes this job.

This fix allows rehersing the job before merges and updates the naming
in multiple places.