ipi/install/rbac: Drop system:unauthenticated group
We have a pull secret to provide authenticated image access.  By
excluding unauthenticated access, we avoid accidentally passing tests
by relying on public image access.  With this commit, we ensure that
the pull secret is getting passed through to the cluster and wired up
to all the image-pulling code that should consume it.