Add tests for s390x
We are using a remote libvirtd instance on a Z machine to run these
tests. Because we don't have a direct connection to the machine, we have
a reverse SSH tunnel connected through the z-bastion, which runs on the
CI cluster. The DNS records are created on the remote machine, but since
we don't use that as our nameserver, we need to artificially add the
records using host aliases. There are also a large number of tests that
are blacklisted in the template. We know these don't work yet, but as
they are fixed, we'll enable them.