Reduce frequency of OpenStack periodic 4.3 and 4.2 jobs
We had to cap the number of simultaneous jobs in order to accomodate
with our overloaded OpenStack CI cloud. This caused many jobs to fail
to get a boskos lease. This commit reduces the frequency of 4.3 and 4.2
periodic jobs to increase the chances of jobs getting a free boskos slot.

Related BZs:
https://bugzilla.redhat.com/show_bug.cgi?id=1788316
https://bugzilla.redhat.com/show_bug.cgi?id=1798939