Overriding build id to prevent libvirt namespace conflict
The libvirt network is currently defined as per the git commit hash. If this job fails to clean up libvirt resources properly, it leaves the environment in an unclean state causing all future jobs to fail because of conflict in defining the libvirt domain. This fix adjusts this by providing the simplest possible unique name - the job number.