Adding openatck upi e2e.
the openstack upi are ientical to ip tests.
According to this comming:
https://github.com/openshift/release/commit/bbd38849

"The openstack e2e job needs some template changes that would break
the aws version of the job. Namely, we need to hardcode a DNS server
that we control, in order to allow the installer to resolve the api
URL.
"

Thus we introduce a new template. If the issues stated in the commit
mentioned are no longer valid. Perhaps the template should be merged
with the other plaform upi-e2e templates.