Adding openatck upi e2e.
It is not possible to use the same
templates for openstack upi as for other platforms
because the upi-installer image used for other
platforms will not work with openstack.
It is missing the openstack client.

For other reasonse reference:
https://github.com/openshift/release/commit/bbd38849

"The openstack e2e job needs some template changes that would break
the aws version of the job. Namely, we need to hardcode a DNS server
that we control, in order to allow the installer to resolve the api
URL.
"