ci-operator: Tar up the must-gather output
The upload sidecar is exploding, with 30 to 40 GB of
container_memory_max_usage_bytes and thousands of goroutines.  That
points to some optimization issues with the sidecar source, but in the
meantime, we expect packing must-gather down into a single file may
reduce sidecar memory usage.  This makes browsing in GCS a bit less
convenient (folks have to pull down the whole tarball instead of
linking to a specific file), but we can revisit the exploded form
if/when we get the sidecar consumption down.

The &&-chaining keeps the raw must-gather directory around in the
event of a failure to create a tarball, because the old-style
templates use 'set +e' in the teardown functions to absorb failures
gathering some logs.  The new step-registry script, on the other hand,
is much more tightly focused and sticks to 'set -o errexit' throughout.