Disable BP only on kubevirt/kubevirt@master
The repo-wide `protect: false` causes OpenShift Prow to attempt to
*disable* BP on all branches on kubevirt/kubevirt, which fails because
our bot does not have rights there. OpenShift Prow is not supposed
to touch BP settings there at all, because another Prow instance is
managing kubevirt BP. OpenShift Prow only attempts to set BP because we
have `protect-tested-repos: true`, and currently we only have kubevirt
jobs on `master`, so it is enough to set `protect: true` on master
branch. This way, branchprotector will not attempt to set BP on other
branches, and it will not attempt to set BP on `master` either, because
`protect: false` has this semantics when set for BP targets which are
only included because of the `protect-nested-repos: true` setting.

Ultimately the branchprotector should be fixed, but such fix will be
quite invasive:

xref: https://github.com/kubernetes/test-infra/issues/14575