Make the OpenStack cloud options configurable
The values used to access the OpenStack cloud (os-cloud, external
network, flavor, etc.) were hardcoded in the containers' `env`
sections.

They are now moved to the top-level `parameters` section and instructs
the `env` sections in the `setup` and `teardown` containers to read
those.

This will enable us to move a subset of our jobs to a different cloud
by setting these parameters in the job templates.