alerting: improve github token exhaustion alert
GitHub does not always restore our token capacity on the hour, or even
hourly. Therefore, we publish the token reset time as well as the
current token capacity. This change makes the alert predict the token
capacity at the time of reset by using the local usage slope, which
improves on the previous alert.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>