ci-operator/step-registry/ipi/conf: Split out per-platform steps
To make it easier to add new platforms, by avoiding a giant platform
switch statement in a single, long configuration file.

Also conditionally add the fips setting.  And switch to tab-indented
Bash for ipi-conf-generic-commands.sh, so I can indent the
here-document payload with <<- [1].  Personally, I'd use tab-intents
for all of this shell, but I've left the four-space indents alone in
the other files to decrease churn.

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_07_04