ci-operator/step-registry: Pivot to explicitly-AWS
And drop all the variant stuff out.  Personally, I prefer a single
chain with a variant knob to make it easy to say "I don't care, pick
something for me" without having to ask for a different workflow.  But
Steve prefers having workflows and jobs tailored to specific
provider/variant combinations.  This commit is an attempt to move in
that direction.  I've dropped the unused refs, but they can be dug
back up out of Git if/when there are jobs that consume them (e.g. GCP
jobs, OVN jobs, pre-existing subnet jobs, etc.).

Shifting the job definitions to their new alphabetical location is
required by the generated-config presubmit [1].

[1]: https://prow.svc.ci.openshift.org/view/gcs/origin-ci-test/pr-logs/pull/openshift_release/7625/pull-ci-openshift-release-master-generated-config/17978