Include ComplianceAsCode/content repo in testing infra
This sets up the necessary files to start testing the
ComplianceAsCode/content repo.

This repo is dedicated on building security content in general. Existing
content already includes RHEL, Fedora, SUSE, OpenStack, and others.
Recently, OpenShift has been included. In order to better build content
for OpenShift, we'd like to start testing the OpenShift-related content
as it's being added.

So this hooks up the needed testing infrastructure so we can test the
content in a live deployment.