ci-operator/step-registry/ipi/deprovision/vsphere: Decompose into sub-steps
So we can recycle the vanilla IPI deprovision step without needing to
make changes to it (other than using the UPI image to get the vSphere
CA certs).  This also makes the link between the resource-creating and
resource-destroying steps more transparent.

Also store the hosted zone ID in a secret so we don't have to look it
up again, which lets us simplify the secret loading a bit in the DNS
deprovisioning.

Remove the ipi-deprovision-vsphere chain, since there's no reason
folks would use it instead of the ipi-vsphere-post chain.