SDCICD-214. Add dynamic version allocation int jobs.
New jobs for integration which select versions based off of the current
production default have been added. As of the time of this commit, these
jobs will add tests for:

- 4.3.0-0.nightly fresh install
- 4.3.0 -> 4.3.0-0.nightly upgrade
- 4.4.0-0.nightly fresh install
- 4.3.0 -> 4.4.0-0.nightly upgrade
- 4.5.0-0.nightly fresh install

When the production default is incremented to the 4.4.z product line,
the y version will be incremented for all of these automatically.
Additionally, once a major version is incremented, that should be
selected as well provided it's within 2 releases of the current
production default.