rehearse: enable rehearsing jobs using config from roe/release
Rehearsals work over in-repo configuration, but our config can be split
between two repos (openshift/release and r-o-e/release). A job in
openshift/release that uses a config for r-o-e/release will fail to be
rehearsed, because its ci-operator config will not be found. We can
solve this by checking out both repos and unifying them.

While it can be argued we should not have such cross-repo jobs, we will
definitely need to have jobs in r-o-e that use steps from a common
registry, and we'll need to rehearse such jobs when their steps change,
which can be solved with the same mechanism.