OpenStack: fix binaries in test image
Newer $IMAGE_TESTS now have /usr/bin/openshift-tests and
/usr/bin/openshift-tests-kubernetes. We need to copy both of them.

Also /usr/bin/kubectl is a symlink pointing to /usr/bin/oc, so unless
the openstack image has the file exists on the OpenStack test image,
this won't work or won't point to the expected version of the binary.
This commit copies all the symlinks pointing to `oc` from the
$IMAGE_TESTS to their new location.

Additionally, copy binaries from $IMAGE_TESTS into a subdirectory in
/tmp/shared to prevent name collision since this directory is prepended
to the PATH and we create files such as `exit` that could shadow real
commands.